Fixed Super-Shrinking Blobs
Added a boolean to keep track of whether a blob was already clicked, and to disable clicking on it if so.

I could have done this better by tracking the blob's state, but I figured just using a boolean would be simpler and just as effective, at least in this specific case.